+++ fe/src/tpl-blogy/components/_homePostsEntryMed.vue
@@ -9,9 +9,9 @@
         <div class="col-sm-6 text-sm-end"><router-link :to="{name: 'Topic', params: {tid: topic.id}}" class="read-more">{{ $t('view_all') }}</router-link></div>
       </div>
       <div class="row">
-        <div class="col-lg-4 mb-4" v-for="doc in docs" v-bind:key="doc.id">
+        <div class="col-lg-4 mb-4" v-for="doc in docs" v-bind:key="doc.id" data-aos="fade-up">
           <div class="post-entry-alt">
-            <router-link :to="{name: 'Document', params: {tid: doc.topic.id, did: doc.id}}" class="img-link">
+            <router-link :to="{name: 'Document', params: {tid: doc.topic.id, did: doc.id}}" class="img-link" data-aos="zoom-in">
               <img :src="$calcDocumentEntryImgUrl(doc, doc.topic.id, '//placehold.co/700x440/214252/90A1A9?text='+$localedText(doc.topic.id).replaceAll(' ','%20'), 'h')" class="img-fluid">
             </router-link>
             <div class="excerpt">
@@ -21,7 +21,7 @@
                 <span class="d-inline-block mt-1">{{ doc.author.name }}</span>
                 <span> - {{ $unixTimestampToReadable(doc.tu) }}</span>
               </div>
-              <p>{{ $localedText(doc.summary) }}</p>
+              <p>{{ $trimText($localedText(doc.summary), 70) }}</p>
               <p><router-link :to="{name: 'Document', params: {tid: doc.topic.id, did: doc.id}}" class="read-more">{{ $t('read') }}</router-link></p>
             </div>
           </div>