+++ fe/src/tpl-blogy/components/_sidebar.vue
@@ -28,9 +28,9 @@
       <h3 class="heading" v-if="$props['documentListTitle']">{{ $props['documentListTitle'] }}</h3>
       <div class="post-entry-sidebar">
         <ul>
-          <li v-for="doc in $props['documentList']" v-bind:key="doc.id">
+          <li v-for="doc in $props['documentList']" v-bind:key="doc.id" data-aos="fade-up">
             <router-link :to="{name: 'Document', params: {tid: doc.topic?doc.topic.id:$props['topic'].id, did: doc.id}}">
-              <img :src="$calcDocumentEntryImgUrl(doc, doc.topic?doc.topic.id:$props['topic'].id, '//placehold.co/440x440/214252/90A1A9?text='+$localedText(doc.id).replaceAll(' ','%20'))" class="me-4 rounded">
+              <img data-aos="zoom-in" :src="$calcDocumentEntryImgUrl(doc, doc.topic?doc.topic.id:$props['topic'].id, '//placehold.co/440x440/214252/90A1A9?text='+$localedText(doc.id).replaceAll(' ','%20'), 's')" class="me-4 rounded">
               <div class="text">
                 <h4>{{ $localedText(doc.title) }}</h4>
                 <div class="post-meta">
@@ -46,7 +46,7 @@
     <div class="sidebar-box">
       <h3 class="heading">{{ $t('topics') }}</h3>
       <ul class="categories">
-        <li v-for="topic in $siteTopics" v-bind:key="topic.id">
+        <li v-for="topic in visibleTopics" v-bind:key="topic.id">
           <router-link :to="{name: 'Topic', params:{tid: topic.id}}">
             <i v-if="topic.icon!=''" :class="topic.icon+' fa-fw'"></i> {{ $localedText(topic.title) }} <span>({{ topic.num_docs }})</span>
           </router-link>
@@ -57,7 +57,7 @@
     <div class="sidebar-box">
       <h3 class="heading">{{ $t('tag_cloud') }}</h3>
       <ul class="tags" style="font-size: small">
-        <li v-for="(_, tag) in $localedText($tagCloud)" v-bind:key="tag">
+        <li v-for="(_, tag) in $localedText($tagCloud)" v-bind:key="tag" :data-aos="randomAOS()" :data-aos-delay="randomNumber(100,1000)" :data-aos-duration="randomNumber(100,1000)">
           <router-link :to="{name: 'TagSearch', query:{q: tag, l: $i18n.locale}}" :class="calcTagCloudCSS(tag)">{{ tag }}</router-link>
         </li>
       </ul>
@@ -70,7 +70,26 @@ export default {
   name: 'lego-page-header',
   inject: ['$global', '$tagCloud', '$siteTopics'],
   props: ['document-list', 'document-list-title', 'topic'],
+  computed: {
+    visibleTopics() {
+      const result = []
+      for (let i = 0; i < this.$siteTopics.length; i++) {
+        if (!this.$siteTopics[i].hidden) {
+          result.push(this.$siteTopics[i])
+        }
+      }
+      return result
+    },
+  },
   methods: {
+    randomNumber(min, max) {
+      return min + Math.floor(Math.random() * (max-min))
+    },
+    randomAOS() {
+      const arr = ['fade-up', 'fade-down', 'fade-right', 'fade-left', 'fade-up-right', 'fade-up-left', 'fade-down-right', 'fade-down-left']
+      const i = this.randomNumber(0, arr.length)
+      return arr[i]
+    },
     calcTagCloudCSS(tag) {
       const cssList = [
         'bg-primary link-light',
@@ -79,7 +98,7 @@ export default {
         'bg-danger link-light',
         'bg-warning text-dark link-dark',
         'bg-info link-light',
-        'bg-light text-dark link-dark',
+        // 'bg-light text-dark link-dark',
       ]
       return this.$pickupFromHash(tag, cssList)
     },