+++ fe/src/tpl-bootstrap/App.vue
@@ -3,6 +3,9 @@
 </template>
 
 <style lang="css">
+/* DevIcons */
+@import "//cdn.jsdelivr.net/gh/devicons/devicon@latest/devicon.min.css";
+
 /* FontAwesome */
 @import "@/_shared/assets/fontawesome-free-6.4.0-web/css/all.min.css";