+++ fe/src/tpl-prettydocs/App.vue
@@ -6,6 +6,9 @@
 /* Google Fonts */
 @import "//fonts.googleapis.com/css?family=Open+Sans:300italic,400italic,600italic,700italic,800italic,400,300,600,700,800";
 
+/* DevIcons */
+@import "//cdn.jsdelivr.net/gh/devicons/devicon@latest/devicon.min.css";
+
 /* FontAwesome */
 @import "@/_shared/assets/fontawesome-free-6.4.0-web/css/all.min.css";